Major update by extracting the section component from the edit page index
